backend/requirements.txt
python-dotenv
pyyaml
funasr
numpy
edge-tts
langchain
langchain-openai
pyads
pyaudio
pynput
colorlog
pygame